test: gate OS-path-specific tests behind //go:build unix (fix Windows build)
TestNewSpecInfoWithCapture_RecordsAllFiles and TestExtract_CrossFileSchemaSlicedFromExternalFile
assert on OS-native $ref path resolution, which uses a different separator on Windows. Move each
into a //go:build unix file so it runs on Linux and macOS but not Windows, instead of comparing
platform-specific paths everywhere.
